requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
pillow
pydantic>=2.0
apscheduler>=3.10.0
